can delete and modify only if user is ok

+++ src/Controller/CustomerController.php
@@ -24,7 +24,10 @@ class CustomerController extends AbstractController
     {
         $user = $this->getUser();
         $paginator->createPagination(Customer::class, ['reseller' => $user], ['createdAt' => "desc"], 'app.customerperpage');
-        return $this->json($paginator->getDatas(), 201, context: ['groups' => 'customer:read']);
+        if($paginator->getDatas() === null){
+            return $this->json('Aucun client trouvé', 404);
+        }
+        return $this->json($paginator->getDatas(), 200, context: ['groups' => 'customer:read']);
     }
 
 
@@ -75,16 +78,26 @@ class CustomerController extends AbstractController
 
     #[Route('/api/customers/{uuid}', name: 'customer_details', methods: ['GET'])]
     public function customerDetails(Uuid $uuid, CustomerRepository $customerRepo): Response
-    {
-        
-        $customer = $customerRepo->findOneBy(['uuid'=>$uuid]);
+    {   
+        //vérification utilisateur (uuid, 404 not found)     
+        $customer = $customerRepo->findOneBy(['uuid'=>$uuid, 'reseller'=>$this->getUser()]);
+        if (!$customer){
+            //doit retouner json
+            return $this->json(["Uuid"=>"Not found"], 404);
+        }
         return $this->json($customer, 201, context: ['groups' => 'customer:read']);
     }
 
     #[Route('/api/customers/{uuid}', name: 'customerModification', methods: ['PUT'])]
-    public function customerModification(Customer $customer, EntityManagerInterface $entityManager, Request $request, SerializerInterface $serializer, ValidatorInterface $validator): Response
+    public function customerModification(Uuid $uuid,CustomerRepository $customerRepo, EntityManagerInterface $entityManager, Request $request, SerializerInterface $serializer, ValidatorInterface $validator): Response
     {
-        
+        $oldCustomer = $customerRepo->findOneBy(['uuid'=>$uuid, 'reseller'=>$this->getUser()]);
+
+        if (!$oldCustomer){
+            //doit retouner json
+            return $this->json(["Uuid"=>"Not found"], 404);
+        }
+
         if (!$request->getContent()) {
             return new Response('
             Le formulaire doit être présenté comme suit:
@@ -96,7 +109,7 @@ class CustomerController extends AbstractController
             }
             ', 400);
         }
-        $customer = $serializer->deserialize($request->getContent(), Customer::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $customer]);
+        $customer = $serializer->deserialize($request->getContent(), Customer::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $oldCustomer]);
 
         $exceptions = $validator->validate($customer);
 
@@ -116,12 +129,14 @@ class CustomerController extends AbstractController
     #[Route('/api/customers/{uuid}', name: 'customerDelete', methods: ['DELETE'])]
     public function customerDelete(Uuid $uuid, CustomerRepository $customerRepository, EntityManagerInterface $em): Response
     {
-        $customer = $customerRepository->findOneBy(['uuid'=> $uuid]);
+        $customer = $customerRepository->findOneBy(['uuid'=> $uuid,  'reseller'=>$this->getUser()]);
+        if (!$customer){
+            //doit retouner json
+            return $this->json(["Uuid"=>"Not found"], 404);
+        }
         $em->remove($customer);
         $em->flush();
-        return new Response('
-        Suppression effectuée
-        ', 200);
+        return $this->json(["User"=>"Deleted"], 200);
         
     }
 }
